move around some stuff
