+++ sigopt/xgboost/run.py
@@ -13,6 +13,7 @@ from .. import create_run
 from ..context import Context
 from ..log_capture import SystemOutputStreamMonitor
 from ..run_context import RunContext
+from .checkpoint_callback import SigOptCheckpointCallback
 from .compute_metrics import compute_classification_metrics, compute_regression_metrics
 
 
@@ -49,8 +50,10 @@ SUPPORTED_OBJECTIVE_PREFIXES = [
 def parse_run_options(run_options):
   if run_options is not None:
     if not isinstance(run_options, dict):
+      # TODO(Harvey): change to actual doc url when it's online
+      doc_url = "https://app.sigopt.com/docs/intro/overview"
       raise ValueError(
-        f"run_options needs to be a dict."
+        f"run_options should be a dictonary. Refer to the sigopt.xgboost.run documentation. {doc_rul}"
       )
 
     if run_options.keys() - DEFAULT_RUN_OPTIONS.keys():
@@ -73,39 +76,6 @@ def parse_run_options(run_options):
     return {**DEFAULT_RUN_OPTIONS, **run_options}
 
   return copy.deepcopy(DEFAULT_RUN_OPTIONS)
-
-
-class SigOptCheckpointCallback(xgboost.callback.TrainingCallback):
-  def __init__(self, run, period=1):
-    self.run = run
-    self.period = period
-    self._latest = None
-    super().__init__()
-
-  def after_iteration(self, model, epoch, evals_log):
-    if not evals_log:
-      return False
-
-    checkpoint_logs = {}
-    for dataset, metric_dict in evals_log.items():
-      for metric_label, metric_record in metric_dict.items():
-        if isinstance(metric_record[-1], tuple):
-          chkpt_value = metric_record[-1][0]
-        else:
-          chkpt_value = metric_record[-1]
-        checkpoint_logs.update({'-'.join((dataset, metric_label)): chkpt_value})
-    if (epoch % self.period) == 0 or self.period == 1:
-      self.run.log_checkpoint(checkpoint_logs)
-      self._latest = None
-    else:
-      self._latest = checkpoint_logs
-
-    return False
-
-  def after_training(self, model):
-    if self._latest is not None:
-      self.run.log_checkpoint(self._latest)
-    return model
 
 
 class XGBRun:
